requirements.txt
Django==4.1.7
django-debug-toolbar==4.4.2
django-paypal==2.0
djangorestframework==3.14.0
paypalrestsdk==1.13.1
Pillow==9.4.0
psycopg2==2.9.9
